Install RVM under the Jenkins Ubuntu user on slaves during bootup.
This is required for the Jenkins Ruby/RVM plugin to function properly.
The plugin is SUPPOSED to be able to do this on its own, but it fails to
properly detect that RVM is missing and install it ... so we need to
install RVM during the bootup process and then let the Ruby/RVM plugin
handle using the installed files as necessary.

Change-Id: I5419d3dc6c4c673eb2553d33ab8673e88450f082