Begin/rescue for deploying the version that's already installed
Added a begin/rescue clause to catch cloning with a server array name
that already exists.  This could happen if a user tries to deploy
a build that's already installed, i.e. we already have server arrays
with that name/version string.  When this happens, we now hide
the traceback and print a friendly message.

Currently the script totally chokes in this scenario.  This change
does not attempt to handle idempotence, it only hides the traceback
and provides a more user-friendly warning.

Change-Id: Ie8c1dc5715e9c8b34c907741f24bdb6cf0310a7c