Begin/rescue for deploying the version that's already installed
Added a begin/rescue clause to catch cloning with a server array name
that already exists.  This could happen if a user tries to deploy
a build that's already installed, i.e. we already have server arrays
with that name/version string.  When this happens, we now hide
the traceback and print a friendly message.

Change-Id: Ie8c1dc5715e9c8b34c907741f24bdb6cf0310a7c