Logic fixes in clone_server_array() method.
1. Set the Puppet inputs first -- in case servers launch the second we
set the array to 'enabled', we want to make sure the puppet inputs are
already set.

2. Don't re-find the server array when we're done. Its unclear why this
was done in the first place, but it seems like its not needed.

3. Don't sleep for 5 seconds after cloning the array -- again, this is
not needed. I did several tests where I hammered the RightScale API in a
loop and never ran into a failure where a sleep would have helped.

Change-Id: I965a8df34f539c5d1a34623d74c6123bde90fd28